pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odysseia-forum"
version = "1.3.0"
description = "Discord论坛搜索机器人"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "discord.py[speed]>=2.3.0",
    "aiohttp>=3.8.0",
    "sqlmodel==0.0.19",
    "aiosqlite>=0.19.0",
    "sqlitefts",
    "rjieba",
    "alembic",
    "fastapi",
    "uvicorn[standard]",
    "uvloop; platform_system != 'Windows'",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "sqlite-utils",
    "ruff"
]

[tool.setuptools.packages.find]
where = ["src"]